requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot
psycopg2-binary
feedparser